Broke out display public data to new module
